+++ docs/changelog.rst
@@ -5,6 +5,7 @@ Version 0.6.0
 -------------
 - Implemented Phenix style sigma A weighting in the Maximum likelihood target (New default for refinement)
 - Added collection versions of the sigma A target
+- Fixed bug in Rfree-generation now min 1000 reflections per bin, min 50 free reflections max 2% and 10 bins
 - Deprecated internal coordinates
 - Renamed Maximum likelihood target to Rice target, Sigma A target to Maximum likelihood target.
 - Moved alignment into experimental